Create basic2.py
Here, we look at While Loops. These types of loops should be used when we don't know exactly when should the loop iteration stop.

For instance, in this example, we are iterating through the positive integers of a sorted list. We don't really know at what location will the positive integers stop. Hence, we use a While loop.
Otherwise, if we knew beforehand that we will have to iterate four times, we would have used a For loop.